all controller completed. gates completed

+++ app/Http/Controllers/CompanyController.php
@@ -5,6 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\company;
 use Illuminate\Http\Request;
 
+
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
+
 class CompanyController extends Controller
 {
     /**
@@ -14,6 +18,8 @@ class CompanyController extends Controller
      */
     public function index()
     {   
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $companies=company::all();
         return view('company.index',compact('companies'));
     }
@@ -24,10 +30,11 @@ class CompanyController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {   
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('company.create');
     }
-
+    
     /**
      * Store a newly created resource in storage.
      *
@@ -37,10 +44,12 @@ class CompanyController extends Controller
     public function store(Request $request)
     {   
         //validation 
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         company::create($request->all()+['users_id'=>auth()->id()]);
-        return redirect()->route('company.index');
+        return redirect()->route('company.index');                        
+        
     }
-
+    
     /**
      * Display the specified resource.
      *
@@ -51,7 +60,7 @@ class CompanyController extends Controller
     {
         //
     }
-
+    
     /**
      * Show the form for editing the specified resource.
      *
@@ -60,9 +69,10 @@ class CompanyController extends Controller
      */
     public function edit(company $company)
     {
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('company.edit',compact('company'));
     }
-
+    
     /**
      * Update the specified resource in storage.
      *
@@ -72,10 +82,11 @@ class CompanyController extends Controller
      */
     public function update(Request $request, company $company)
     {
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $company->update($request->all()+['users_id'=>auth()->id()]);
         return redirect()->route('company.index');
     }
-
+    
     /**
      * Remove the specified resource from storage.
      *
@@ -84,6 +95,7 @@ class CompanyController extends Controller
      */
     public function destroy(company $company )
     {
+        abort_if(Gate::denies('company_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $company->delete();
         return redirect()->route('company.index');
     }
